mito-ai: create base websocket client
